fix: 클래스 세분화 및 파일 재배치
