+++ mobile_app/lib/pages/main_page.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:mobile_app/api/api.dart';
-import 'package:mobile_app/cafe_image_slider.dart';
 import 'package:mobile_app/constants/color.dart';
 import 'package:mobile_app/constants/type.dart';
-import 'package:mobile_app/header.dart';
-import 'package:mobile_app/main_button.dart';
-import 'package:mobile_app/main_slider_section.dart';
-import 'package:mobile_app/main_tab.dart';
-import 'package:mobile_app/main_table_section.dart';
+import 'package:mobile_app/view/cafe_main/main_button.dart';
+import 'package:mobile_app/view/cafe_main/main_slider_section.dart';
+import 'package:mobile_app/view/cafe_main/main_tab.dart';
+import 'package:mobile_app/view/cafe_main/main_table_section.dart';
+import 'package:mobile_app/view/common/header.dart';
+import 'package:mobile_app/view/common/image_index_bullet.dart';
 
 class MainScreen extends StatefulWidget {
   @override
