+++ mobile_app/lib/pages/settings_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mobile_app/header.dart';
+import 'package:mobile_app/view/common/header.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
 class SettingsScreen extends StatelessWidget {
