+++ mobile_app/lib/view/cafe_detail/cafe_detail_info.dart
@@ -1,42 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:mobile_app/cafe_info.dart';
 import 'package:mobile_app/constants/type.dart';
-import 'package:mobile_app/constants/util.dart';
+import 'package:mobile_app/util/cafe_detail.dart';
+import 'package:mobile_app/util/common.dart';
+import 'package:mobile_app/view/common/cafe_name.dart';
 
-class CafeMainInfo extends StatelessWidget {
-  final CafeModel cafe;
-
-  CafeMainInfo({required this.cafe});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-        padding: EdgeInsets.only(left: 20, right: 20, bottom: 8),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            CafeName(name: cafe.name),
-            SizedBox(height: 4),
-            Text(
-              '${cafe.place.name} OPEN ${cafe.metadata?.hour ?? ''}',
-              style: TextStyle(fontSize: 14),
-            ),
-            SizedBox(
-              height: 4,
-            ),
-            Container(
-              alignment: Alignment.centerRight,
-              child: Text(
-                "${cafe.metadata?.creator ?? 'jyuunnii'} 님이 올려주신 ${cafe.name}",
-                style: TextStyle(
-                  fontSize: 11,
-                ),
-              ),
-            )
-          ],
-        ));
-  }
-}
+import 'cafe_detail_info_item.dart';
+import 'cafe_detail_info_subway_item.dart';
 
 class CafeDetailInfo extends StatelessWidget {
   final CafeModel cafe;
@@ -94,20 +63,6 @@ class CafeDetailInfo extends StatelessWidget {
           onPressed: () => handleInstagramClick(cafe.name, context),
         )
       ],
-    );
-  }
-}
-
-class CafeName extends StatelessWidget {
-  final String name;
-
-  CafeName({required this.name});
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      name,
-      style: TextStyle(fontSize: 14, fontWeight: FontWeight.bold),
     );
   }
 }
