+++ mobile_app/lib/view/cafe_main/main_bottom_sheet.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:mobile_app/api/api.dart';
-import 'package:mobile_app/cafe_image.dart';
 import 'package:mobile_app/constants/color.dart';
 import 'package:mobile_app/constants/type.dart';
 import 'package:mobile_app/router/mixins/enter_cafe_detail_mixin.dart';
+
+import 'cafe_image_card.dart';
 
 class MainBottomSheet extends StatefulWidget {
   final List<CafeModel> hotCafeList;
@@ -91,7 +92,7 @@ class _MainBottomSheetState extends State<MainBottomSheet>
                                         right: index == 9 ? 20 : 4,
                                         left: index == 0 ? 20 : 0),
                                     child: GestureDetector(
-                                      child: RepresentativeCafe(
+                                      child: CafeImageCard(
                                           cafe: (_hotCafeList ??
                                               widget.hotCafeList)[index]),
                                       onTap: enterDetail((_hotCafeList ??
@@ -142,41 +143,5 @@ class _MainBottomSheetState extends State<MainBottomSheet>
                 ))
           ],
         ));
-  }
-}
-
-class RepresentativeCafe extends StatelessWidget {
-  final CafeModel cafe;
-
-  RepresentativeCafe({required this.cafe});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Card(
-            clipBehavior: Clip.antiAlias,
-            elevation: 0,
-            margin: EdgeInsets.only(right: 3),
-            child: Container(
-              width: 100,
-              height: 100,
-              decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(4),
-                  color: Palette.grayBG),
-              child: CafeImage(image: cafe.image.mainImage, size: 100),
-            )),
-        Container(
-            width: 100,
-            margin: EdgeInsets.only(top: 8),
-            child: Text(cafe.name,
-                overflow: TextOverflow.ellipsis,
-                style: TextStyle(fontSize: 12))),
-        SizedBox(height: 2),
-        Text(cafe.place.name,
-            style: TextStyle(fontSize: 11, color: Palette.gray))
-      ],
-    );
   }
 }
