+++ mobile_app/lib/view/cafe_main/main_button.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:mobile_app/constants/color.dart';
 import 'package:mobile_app/constants/type.dart';
-import 'package:mobile_app/constants/util.dart';
-import 'package:mobile_app/main_bottom_sheet.dart';
+import 'package:mobile_app/util/common.dart';
+import 'package:mobile_app/view/cafe_main/main_bottom_sheet.dart';
 
 class MainButtonSetOfSlider extends StatelessWidget {
   final List<CafeModel> hotCafeList;
