+++ mobile_app/lib/view/common/cafe_name.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class CafeName extends StatelessWidget {
+  final String name;
+
+  CafeName({required this.name});
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(
+      name,
+      style: TextStyle(fontSize: 14, fontWeight: FontWeight.bold),
+    );
+  }
+}
