+++ mobile_app/lib/view/common/skeleton.dart
